[Feat] Calendar 컴포넌트 삭제
